+++ parser-typechecker/unison-parser-typechecker.cabal
@@ -64,8 +64,6 @@ library
     Unison.Typechecker.Context
     Unison.Typechecker.Components
     Unison.Typechecker.Extractor
-    Unison.Typechecker.Extractor2
-    Unison.Typechecker.Extractor3
     Unison.Typechecker.TypeError
     Unison.UnisonFile
     Unison.Util.AnnotatedText
